created directory form and moved Components with common styles inside
