+++ src/components/form/FormTextInput/FormTextInput.module.css
@@ -1,4 +1,4 @@
-@import '../../utilities/styling/mixins/typography.css';
+@import '../../../utilities/styling/mixins/typography.css';
 
 .inputThumb {
     display: flex;
